[NHDR-211] refactor: IHT -> IHL 파일명 변경
