+++ src/pages/inheritance/IHL4_InheritanceWillPage.vue
