+++ src/pages/inheritance/IHL5_InheritanceResultPage.vue
